Controller for store methods
