+++ student-store-express-api/app.js
@@ -14,32 +14,6 @@ app.get("/", (_req, res) => {
     })
 })
 
-app.get("/store", (_req, res) => {
-    res.status(200).json({
-        products: Store.listProducts()
-    })
-})
-
-app.get("/store/:productId", (req, res) => {
-    const {productId} = req.params;
-    res.json({
-        product: Store.fetchProduct(productId)
-    })
-})
-
-app.post("/store", (req, res) => {
-    const { shoppingCart, user } = req.body;
-
-    const purchase = Store.createOrder( {
-        user,
-        shoppingCart
-    } );
-
-    res.status(201).json({
-        purchase
-    })
-})
-
 // Error handlers
 
 app.use((error, _req, res, _next) => {
